pvr.dvbviewer: 64bit channel ids, code cleanup and much more
this commit is gone bad as it contains too much changes. at first I
tried to do just some code cleanup, but as I jumped from block to block
I just found more and more positions to cleanup. after a few hours of
work it was impossible to follow the changes.

this commit includes:
* massive code cleanup
* in generell more C++-ish code
* use iterator instead of index access where possible
* read, calculate and use 64bit channel ids
* for favourites there's a fallback for old 32bit channel ids
* use channel names from favourites
* replace "/"-character with " " in channel logo urls
* resort methods and member variables in DvbData.h
* resort methods in DvbData.cpp to match the new sorting of DvbData.h
* less LOG_INFO logs, better LOG_DEBUG logs
* use PRId64/PRIu64 for int64_t/uint64_t
* probably much more